Imported PopupDisplay and Created toString(). Then added a for loop to display the toString() text and display all the different Kahoots.
Created and imported the popupDisplay as well as made a toString() in the Kahoot class then added it to the controller. Then created a for loop and added instances of the helper methods to have the program go through all the Kahoots made and display it's name, who it was made by, and how many questions it had in the GUI using popup.displayText.