Added images of certain sections of the Controller file to keep as notes.
Also fixed a typo in the PopupDisplay.java file. Also note: the // parts on the //Kahoot currentKahoot = myKahoots.get(index); and the line below are not commented out in the program due to an error in another part (a for loop i believe?) that used "creator" and would've had to use "currentCreator" instead.