Added a new method for a section on changing parts of the list. Then created another one to practice using those change on our own.
Two more methods were created and added to the code's start. The first one called changeTheList() was done as a class to cover the changes while listPractice() was made to use those changes. listPractice also uses for loops to display the list forwards, and then backwards. Also commented out 2 for loops that displayed the kahoot names letter by letter for easier reading when running the program as it is not needed.